Various tweaks
I moved several classes out into their own files and re-did the FormerHumanChance and AlwaysFormerHuman classes to improve readablility and reuseability.
