+++ Source/Pawnmorphs/Esoteria/CompProperties_AlwaysFormerHuman.cs
@@ -1,0 +1,14 @@
+﻿using Verse;
+
+namespace Pawnmorph
+{
+    public class CompProperties_AlwaysFormerHuman : CompProperties
+    {
+        public HediffDef hediff;
+
+        public CompProperties_AlwaysFormerHuman()
+        {
+            compClass = typeof(CompAlwaysFormerHuman);
+        }
+    }
+}
