+++ Source/Pawnmorphs/Esoteria/HediffCompProperties_Single.cs
@@ -1,0 +1,12 @@
+﻿using Verse;
+
+namespace Pawnmorph
+{
+    public class HediffCompProperties_Single : HediffCompProperties
+    {
+        public HediffCompProperties_Single()
+        {
+            compClass = typeof(HediffComp_Single);
+        }
+    }
+}
