+++ Source/Pawnmorphs/Esoteria/HediffCompProperties_TerrainBasedMorph.cs
@@ -1,0 +1,15 @@
+﻿using Verse;
+
+namespace Pawnmorph
+{
+    public class HediffCompProperties_TerrainBasedMorph : HediffCompProperties
+    {
+        public HediffDef hediffDef = null;
+        public TerrainDef terrain = null;
+
+        public HediffCompProperties_TerrainBasedMorph()
+        {
+            compClass = typeof(HediffComp_TerrainBasedMorph);
+        }
+    }
+}
