+++ Source/Pawnmorphs/Esoteria/HediffComp_Single.cs
@@ -1,0 +1,15 @@
+﻿using Verse;
+
+namespace Pawnmorph
+{
+    public class HediffComp_Single : HediffComp
+    {
+        public HediffCompProperties_Single Props
+        {
+            get
+            {
+                return (HediffCompProperties_Single)this.props;
+            }
+        }
+    }
+}
