+++ Source/Pawnmorphs/Esoteria/HediffComp_Staged.cs
@@ -1,0 +1,15 @@
+﻿using RimWorld;
+
+
+namespace Pawnmorph
+{
+    public class HediffComp_Staged
+    {
+        public float daysToProduce = 1;
+        public int amount = 1;
+        public string resource = "Chemfuel";
+        public float chance = 50;
+        public string rareResource = "Chemfuel";
+        public ThoughtDef thought = null;
+    }
+}
