+++ Source/Pawnmorphs/Esoteria/HediffGiver_PermanentFeral.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using Verse;
+using RimWorld;
+
+namespace Pawnmorph
+{
+    public class HediffGiver_PermanentFeral : HediffGiver
+    {
+        public float mtbDays;
+
+        public override void OnIntervalPassed(Pawn pawn, Hediff cause)
+        {
+            if (base.TryApply(pawn, null))
+            {
+                Find.LetterStack.ReceiveLetter("LetterHediffFromPermanentTFLabel".Translate(pawn.LabelShort).CapitalizeFirst(), "LetterHediffFromPermanentTF".Translate(pawn.LabelShort).CapitalizeFirst(), LetterDefOf.NegativeEvent, pawn, null, null);
+            }
+        }
+    }
+}
