+++ Source/Pawnmorphs/Esoteria/PawnmorphHediffGiverUtility.cs
@@ -1,0 +1,69 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using Verse;
+using RimWorld;
+
+namespace Pawnmorph
+{
+    public static class PawnmorphHediffGiverUtility
+    {
+        public static bool TryApply(Pawn pawn, HediffDef hediff, List<BodyPartDef> partsToAffect = null, bool canAffectAnyLivePart = false, int countToAffect = 1, List<Hediff> outAddedHediffs = null)
+        {
+            try
+            {
+                if (canAffectAnyLivePart || partsToAffect != null)
+                {
+                    bool result = false;
+                    for (int i = 0; i < countToAffect; i++)
+                    {
+
+                        IEnumerable<BodyPartRecord> source = pawn.health.hediffSet.GetNotMissingParts(BodyPartHeight.Undefined, BodyPartDepth.Undefined, null, null);
+                        if (partsToAffect != null)
+                        {
+                            source = from p in source
+                                     where partsToAffect.Contains(p.def)
+                                     select p;
+                        }
+                        if (canAffectAnyLivePart)
+                        {
+                            source = from p in source
+                                     where p.def.alive
+                                     select p;
+                        }
+                        source = from p in source
+                                 where !pawn.health.hediffSet.HasHediff(hediff, p, false) && !pawn.health.hediffSet.PartOrAnyAncestorHasDirectlyAddedParts(p)
+                                 select p;
+                        if (!source.Any<BodyPartRecord>())
+                        {
+                            break;
+                        }
+
+                        BodyPartRecord partRecord = source.First();
+                        Hediff hediff2 = HediffMaker.MakeHediff(hediff, pawn, partRecord);
+                        pawn.health.AddHediff(hediff2, null, null, null);
+                        if (outAddedHediffs != null)
+                        {
+                            outAddedHediffs.Add(hediff2);
+                        }
+                        result = true;
+                    }
+                    return result;
+                }
+                if (!pawn.health.hediffSet.HasHediff(hediff, false))
+                {
+                    Hediff hediff3 = HediffMaker.MakeHediff(hediff, pawn, null);
+                    pawn.health.AddHediff(hediff3, null, null, null);
+                    if (outAddedHediffs != null)
+                    {
+                        outAddedHediffs.Add(hediff3);
+                    }
+                    return true;
+                }
+                return false;
+            }
+            catch { return false; }
+        }
+    }
+}
